Refactor: modern naming conventions, bug fixes, and Author CRUD

+++ Controllers/BooksController.cs
@@ -1,12 +1,10 @@
-﻿using AprilBookStore.DataAccess;
+using AprilBookStore.DataAccess;
 using AprilBookStore.Models;
 using AprilBookStore.ViewModels.Book;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using System.Collections.ObjectModel;
 using System.Net;
-using System.Text.RegularExpressions;
 using X.PagedList;
 
 namespace AprilBookStore.Controllers
@@ -14,26 +12,26 @@ namespace AprilBookStore.Controllers
     [Authorize(Roles = "Admin,SuperAdmin")]
     public class BooksController : Controller
     {
-        private readonly IData data;
-        private readonly IWebHostEnvironment env;
+        private readonly IData _data;
+        private readonly IWebHostEnvironment _environment;
 
-        public BooksController(IData data, IWebHostEnvironment hostEnvironment)
+        public BooksController(IData data, IWebHostEnvironment environment)
         {
-            this.data = data;
-            this.env = hostEnvironment;
+            _data = data;
+            _environment = environment;
         }
 
         [AcceptVerbs("GET", "POST")]
         [AllowAnonymous]
         public async Task<IActionResult> AutoComplete(string term)
         {
-            var result = (await data.SearchBook(term)).Select(b => b.Name).ToList();
+            var result = (await _data.SearchBook(term)).Select(b => b.Name).ToList();
             return Json(result);
         }
 
-        public async Task<IActionResult> Name(string Name)
+        public async Task<IActionResult> Name(string name)
         {
-            if ((await data.SearchBook(Name)).Any(b => b.Name == Name))
+            if ((await _data.SearchBook(name)).Any(b => b.Name == name))
             {
                 return Json(false);
             }
@@ -42,23 +40,22 @@ namespace AprilBookStore.Controllers
 
         public ActionResult Index(int? page)
         {
-            var books = data.GetBooks().OrderBy(b => b.Name);
+            var books = _data.GetBooks().OrderBy(b => b.Name);
             return View(books.ToPagedList(page ?? 1, 25));
         }
 
         [HttpPost]
-       public async Task<ActionResult> Index(string name, int? categoryId, decimal? minPrice, decimal? maxPrice, int? page)
+        public async Task<ActionResult> Index(string name, int? categoryId, decimal? minPrice, decimal? maxPrice, int? page)
         {
-
             ICollection<Book> books;
 
             if (!string.IsNullOrEmpty(name))
             {
-                 books = await data.SearchBook(name);
+                books = await _data.SearchBook(name);
             }
             else
             {
-                books = data.GetBooks().OrderBy(b => b.Name).ToList();
+                books = _data.GetBooks().OrderBy(b => b.Name).ToList();
             }
 
             if (categoryId.HasValue)
@@ -81,17 +78,18 @@ namespace AprilBookStore.Controllers
             ViewData["MinPrice"] = minPrice;
             ViewData["MaxPrice"] = maxPrice;
 
-            ViewBag.Categories = new SelectList(data.GetCategories(), "Id", "Name");
+            ViewBag.Categories = new SelectList(_data.GetCategories(), "Id", "Name");
 
             int pageSize = 25;
-            int pageNumber = (page ?? 1);
+            int pageNumber = page ?? 1;
             return View(await books.ToPagedListAsync(pageNumber, pageSize));
         }
 
         // GET: Books/Details/5
+        [AllowAnonymous]
         public ActionResult Details(int id)
         {
-            Book? book = data.GetBook(id);
+            Book? book = _data.GetBook(id);
             if (book == null)
             {
                 return View("NotFound");
@@ -102,8 +100,8 @@ namespace AprilBookStore.Controllers
         // GET: Books/Create
         public ActionResult Create()
         {
-            ViewBag.AuthorId = new SelectList(data.GetAuthors(), "Id", "Name");
-            ViewBag.CategoryId = new SelectList(data.GetCategories(), "Id", "Name");
+            ViewBag.AuthorId = new SelectList(_data.GetAuthors(), "Id", "Name");
+            ViewBag.CategoryId = new SelectList(_data.GetCategories(), "Id", "Name");
             return View();
         }
 
@@ -134,7 +132,7 @@ namespace AprilBookStore.Controllers
                 {
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + book.BookCoverFile.FileName;
                     string imagePath = Path.Combine("/book-covers/UploadedCovers", uniqueFileName);
-                    string filePath = Path.Combine(env.ContentRootPath + "\\wwwroot\\book-covers\\UploadedCovers", uniqueFileName);
+                    string filePath = Path.Combine(_environment.WebRootPath, "book-covers", "UploadedCovers", uniqueFileName);
                     using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
                         book.BookCoverFile.CopyTo(fileStream);
@@ -142,12 +140,12 @@ namespace AprilBookStore.Controllers
                     newBook.ImgPath = imagePath;
                 }
 
-                data.AddBook(newBook);
+                _data.AddBook(newBook);
                 return RedirectToAction("Index");
             }
 
-            ViewBag.AuthorId = new SelectList(data.GetAuthors(), "Id", "Name", book.AuthorId);
-            ViewBag.CategoryId = new SelectList(data.GetCategories(), "Id", "Name", book.CategoryId);
+            ViewBag.AuthorId = new SelectList(_data.GetAuthors(), "Id", "Name", book.AuthorId);
+            ViewBag.CategoryId = new SelectList(_data.GetCategories(), "Id", "Name", book.CategoryId);
             return View(book);
         }
 
@@ -159,7 +157,7 @@ namespace AprilBookStore.Controllers
                 return new StatusCodeResult((int)HttpStatusCode.BadRequest);
             }
 
-            Book? book = data.GetBook(id.Value);
+            Book? book = _data.GetBook(id.Value);
 
             if (book == null)
             {
@@ -171,20 +169,20 @@ namespace AprilBookStore.Controllers
                 Id = id.Value,
                 AuthorId = book.AuthorId,
                 CategoryId = book.CategoryId,
-                BookStar =book.BookStar??0,
+                BookStar = book.BookStar ?? 0,
                 Description = book.Description,
                 Format = book.Format,
-                ISBN=book.ISBN,
+                ISBN = book.ISBN,
                 QuantityInStock = book.QuantityInStock,
-                PublicationYear=book.PublicationYear,
+                PublicationYear = book.PublicationYear,
                 Name = book.Name,
-                Price=book.Price,
-                IsDeleted= book.IsDeleted,
-                IsVisible= book.IsVisible,
+                Price = book.Price,
+                IsDeleted = book.IsDeleted,
+                IsVisible = book.IsVisible,
                 CreatedDate = book.CreatedDate
             };
-            ViewBag.AuthorId = new SelectList(data.GetAuthors(), "Id", "Name", book.AuthorId);
-            ViewBag.CategoryId = new SelectList(data.GetCategories(), "Id", "Name", book.CategoryId);
+            ViewBag.AuthorId = new SelectList(_data.GetAuthors(), "Id", "Name", book.AuthorId);
+            ViewBag.CategoryId = new SelectList(_data.GetCategories(), "Id", "Name", book.CategoryId);
             return View(viewModel);
         }
 
@@ -192,20 +190,19 @@ namespace AprilBookStore.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(EditBookViewModel editedBook)
         {
-            var book = data.GetBook(editedBook.Id);
+            var book = _data.GetBook(editedBook.Id);
             if (book == null)
             {
                 return NotFound();
             }
-            if (!((await data.SearchBook(editedBook.Name)).Any(b => b.Name == editedBook.Name) && editedBook.Name!=book.Name))
+            if (!((await _data.SearchBook(editedBook.Name)).Any(b => b.Name == editedBook.Name) && editedBook.Name != book.Name))
             {
-
                 book.Name = editedBook.Name;
                 book.Format = editedBook.Format;
                 book.ISBN = editedBook.ISBN;
                 book.Price = editedBook.Price;
-                book.IsDeleted= editedBook.IsDeleted;
-                book.IsVisible=editedBook.IsVisible;
+                book.IsDeleted = editedBook.IsDeleted;
+                book.IsVisible = editedBook.IsVisible;
                 book.PublicationYear = editedBook.PublicationYear;
                 book.QuantityInStock = editedBook.QuantityInStock;
                 book.AuthorId = editedBook.AuthorId;
@@ -214,23 +211,28 @@ namespace AprilBookStore.Controllers
 
                 if (editedBook.BookCoverFile != null && editedBook.BookCoverFile.Length > 0)
                 {
+                    if (!string.IsNullOrEmpty(book.ImgPath))
+                    {
+                        string oldFilePath = Path.Combine(_environment.WebRootPath, book.ImgPath.TrimStart('/', '\\'));
+                        if (System.IO.File.Exists(oldFilePath))
+                        {
+                            System.IO.File.Delete(oldFilePath);
+                        }
+                    }
+
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + editedBook.BookCoverFile.FileName;
                     string imagePath = Path.Combine("/book-covers/UploadedCovers", uniqueFileName);
-                    string filePath = Path.Combine(env.ContentRootPath + "\\wwwroot\\book-covers\\UploadedCovers", uniqueFileName);
+                    string filePath = Path.Combine(_environment.WebRootPath, "book-covers", "UploadedCovers", uniqueFileName);
                     using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
                         editedBook.BookCoverFile.CopyTo(fileStream);
-                    }
-                    if (System.IO.File.Exists(filePath))
-                    {
-                        System.IO.File.Delete(filePath);
                     }
                     book.ImgPath = imagePath;
                 }
 
                 if (ModelState.IsValid)
                 {
-                    data.UpdateBook(book);
+                    _data.UpdateBook(book);
                     return RedirectToAction("Index");
                 }
             }
@@ -238,8 +240,8 @@ namespace AprilBookStore.Controllers
             {
                 ModelState.AddModelError("Name", "This name is used !!");
             }
-            ViewBag.AuthorId = new SelectList(data.GetAuthors(), "Id", "Name", book.AuthorId);
-            ViewBag.CategoryId = new SelectList(data.GetCategories(), "Id", "Name", book.CategoryId);
+            ViewBag.AuthorId = new SelectList(_data.GetAuthors(), "Id", "Name", book.AuthorId);
+            ViewBag.CategoryId = new SelectList(_data.GetCategories(), "Id", "Name", book.CategoryId);
             return View(editedBook);
         }
 
@@ -251,7 +253,7 @@ namespace AprilBookStore.Controllers
                 return new StatusCodeResult((int)HttpStatusCode.BadRequest);
             }
 
-            Book book = data.GetBook(id.Value);
+            Book? book = _data.GetBook(id.Value);
             if (book == null)
             {
                 return NotFound();
@@ -264,25 +266,27 @@ namespace AprilBookStore.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Book? book = data.GetBook(id);
+            Book? book = _data.GetBook(id);
             if (book == null)
             {
                 return NotFound();
             }
 
-            data.DeleteBook(book.Id);
+            _data.DeleteBook(book.Id);
             try
             {
-                string fileName = Path.GetFileName(book.ImgPath.Split('/').Last());
-                string filePath = Path.Combine(env.WebRootPath, "book-covers/UploadedCovers", fileName);
-                if (System.IO.File.Exists(filePath))
+                if (!string.IsNullOrEmpty(book.ImgPath))
                 {
-                    System.IO.File.Delete(filePath);
+                    string fileName = Path.GetFileName(book.ImgPath.Split('/').Last());
+                    string filePath = Path.Combine(_environment.WebRootPath, "book-covers", "UploadedCovers", fileName);
+                    if (System.IO.File.Exists(filePath))
+                    {
+                        System.IO.File.Delete(filePath);
+                    }
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-
             }
 
             return RedirectToAction("Index");
@@ -292,7 +296,7 @@ namespace AprilBookStore.Controllers
         {
             if (disposing)
             {
-                data.Dispose();
+                _data.Dispose();
             }
             base.Dispose(disposing);
         }
